move : checkicon to progresscheckicon
