+++ src/assets/ProgressCheckIcon.tsx
